fixed the float problem that wanted to be an integer
the float didn't work so fixed it and made sure it was working without errors

+++ Pokemon.cs
@@ -43,7 +43,7 @@ namespace Pokebattle
             }
             foreach(var weakness in Weaknesses)
             {
-                if (weakness.Type == type) return weakness.Multiplier;
+                if (weakness.Type == type) return (int)weakness.Multiplier;
             }
 
             return 1;
